refactor: move agent code to src/agent/ with clean module boundary
Move all agent/worker files into src/agent/ to mirror the foreman's
src/foreman/ structure. repl.ts becomes agent/index.ts as the entry
point. Internal cross-imports stay as ./ references; imports to shared
files (config, types, utils) become ../ references. Update package.json
scripts, bin/brunel.cjs, all test imports, and CLAUDE.md documentation.

Closes #509
