+++ package.json
@@ -9,8 +9,8 @@
     "build": "cd frontend && npm install && npm run build",
     "start": "tsx src/foreman/index.ts",
     "dev": "tsx watch src/foreman/index.ts",
-    "repl": "tsx src/repl.ts",
-    "worker": "tsx src/repl.ts --worker-mode",
+    "repl": "tsx src/agent/index.ts",
+    "worker": "tsx src/agent/index.ts --worker-mode",
     "test": "vitest run",
     "test:watch": "vitest",
     "smoke": "tsx tests/smoke.ts",
