+++ src/agent/input.ts
