+++ src/agent/model.ts
